pyproject.toml
[tool.poetry]
name = "posthub-backend"
version = "0.1.0"
description = "Backend, Startup Structura"
authors = ["yepIwt <ReetRay@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
alembic = "^1.12.1"
pre-commit = "^3.5.0"
uvicorn = "^0.24.0"
fastapi = "^0.104.1"
orjson = "^3.9.10"
contextvars = "^2.4"
pydantic = "^1.10.7"
asyncpg = "^0.29.0"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.9"
celery = "^5.3.5"
vk-api = "^11.9.9"
python-telegram-bot = "^20.6"
pyjwt = "^2.8.0"
fastapi-users = {extras = ["sqlalchemy"], version = "^12.1.2"}
guardpost = {extras = ["jwt"], version = "^1.0.2"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
